component/StatisticOne : fixing bug with loading and editing arrow
